pyproject.toml
[tool.poetry]
name = "macabake"
version = "0.0.1"
description = ""
authors = ["Sam Korn <korn94sam@gmail.com>"]

packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
lnd-grpc-client = "^0.4.4"
click = "^8.1.3"

#[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
macabake = "src:macabake"